chore: clean up code
- Improve styles
- Remove unnecessary code
