+++ src/app/dashboard/position-datas/page.tsx
@@ -4,6 +4,7 @@ import { getPositionDatasByProjectId } from "@/lib/services"
 import { TablePositionDatas } from "@/ui/dashboard/postion-datas/table-position-datas"
 import { AddNewPositionData } from "@/ui/dashboard/postion-datas/add-new-position-datas"
 import { getCookieToken } from "@/lib/services/cookies"
+import { redirect } from "next/navigation"
 
 export const metadata: Metadata = {
     title: "Position Datas",
@@ -12,6 +13,7 @@ export const metadata: Metadata = {
 
 const getInformation = async () => {
     const { ok, data } = await getCookieToken()
+    if (!ok) return redirect("/dashboard")
     return await getPositionDatasByProjectId(data.idProject)
 }
