+++ src/app/dashboard/position-soil-datas/[position-soil-dataId]/page.tsx
@@ -2,7 +2,8 @@ import Link from "next/link"
 import { Metadata } from "next"
 import { Params } from "@/lib/@types/types"
 import { getPositionSoilDataById } from "@/lib/services/position-soil-datas"
-import { camelCaseToWords, evalutionGrosor } from "@/lib/utils"
+import { camelCaseToWords } from "@/lib/utils"
+import { evalutionGrosor } from "@/lib/math"
 import { Button } from "@halvaradop/ui-button"
 import { isObject } from "@halvaradop/ts-utility-types/validate"
