--- src/app/dashboard/position-soil-datas/add/page.tsx
@@ -1,13 +1,0 @@
-import { Metadata } from "next"
-import { AddPositionSoilDatas } from "@/ui/dashboard/position-soil-datas/add-position-soil-datas"
-
-export const metadata: Metadata = {
-    title: "Add sample",
-    description: "Add a sample to the dashboard",
-}
-
-const AddPositionSoilDatasPage = async () => {
-    return <AddPositionSoilDatas />
-}
-
-export default AddPositionSoilDatasPage
