+++ src/ui/common/menu-routes.tsx
@@ -1,64 +1,73 @@
 import Link from "next/link"
-import { camelCaseToHyphenCamel, camelCaseToWords, merge } from "@/lib/utils"
 import { MenuRoutesProps } from "@/lib/@types/props"
 import { roleBasedAccessControl } from "@/middleware"
 import { headerMenuListVariants } from "@/ui/motion/header-menu.motion"
 import * as motion from "framer-motion/client"
+import { merge } from "@halvaradop/ui-core"
 
-const links = {
-    projects: [
-        { href: "/dashboard/projects", label: "List" },
-        { href: "/dashboard/projects/add", label: "Add" },
-    ],
-    stakeHolders: [
-        { href: "/dashboard/stake-holders", label: "List" },
-        { href: "/dashboard/stake-holders/add", label: "Add" },
-    ],
-    contactPeople: [
-        { href: "/dashboard/contact-people", label: "List" },
-        { href: "/dashboard/contact-people/add", label: "Add" },
-    ],
-    fields: [
-        { href: "/dashboard/fields", label: "List" },
-        { href: "/dashboard/fields/add", label: "Add" },
-    ],
-    addresses: [
-        { href: "/dashboard/addresses", label: "List" },
-        { href: "/dashboard/addresses/add", label: "Add" },
-    ],
-    positionDatas: [
-        { href: "/dashboard/position-datas", label: "List" },
-        { href: "/dashboard/position-datas/add", label: "Add" },
-    ],
-    positionSoilDatas: [
-        { href: "/dashboard/position-soil-datas", label: "List" },
-        { href: "/dashboard/position-soil-datas/add", label: "Add" },
-    ],
-}
+const links = [
+    {
+        title: "Projects",
+        link: "/dashboard/projects",
+    },
+    {
+        title: "Stakeholders",
+        link: "/dashboard/stake-holders",
+    },
+    {
+        title: "Contact People",
+        link: "/dashboard/contact-people",
+    },
+    {
+        title: "Fields",
+        link: "/dashboard/fields",
+    },
+    {
+        title: "Addresses",
+        link: "/dashboard/addresses",
+    },
+    {
+        title: "Position Datas",
+        link: "/dashboard/position-datas",
+    },
+    {
+        title: "Position Soil Datas",
+        link: "/dashboard/position-soil-datas",
+    },
+]
 
-export const MenuRoutes = ({ className, classTitle, classOption, session }: MenuRoutesProps) => {
+export const MenuRoutes = ({ className, classNameTitle, classNameOption, session }: MenuRoutesProps) => {
     if (!session) return null
     const {
         user: { role },
     } = session
     const rbac = roleBasedAccessControl[role] ?? []
 
-    return Object.entries(links).map(([key, links]) => {
-        if (!rbac.includes(camelCaseToHyphenCamel(key))) return null
-
-        return (
-            <motion.ul className={merge("space-y-1", className)} variants={headerMenuListVariants} key={key}>
-                <li className={merge("font-medium capitalize", classTitle)}>{camelCaseToWords(key)}</li>
-                {links.map(({ href, label }, key) => {
-                    const pathname = href.replace(/^\/dashboard\/?/, "")
-                    if (!rbac.includes(camelCaseToHyphenCamel(pathname)) && pathname !== "") return null
-                    return (
-                        <li className={merge("ml-3", classOption)} key={key}>
-                            <Link href={href}>{label}</Link>
-                        </li>
-                    )
-                })}
-            </motion.ul>
-        )
-    })
+    return (
+        <motion.ul className={merge("w-full", className)} variants={headerMenuListVariants}>
+            <li
+                className={merge(
+                    "w-full p-2 font-medium border border-transparent rounded hover:text-white hover:border-sky-500 hover:cursor-pointer hover:bg-sky-500",
+                    classNameTitle
+                )}
+            >
+                <Link href="/dashboard">Overview</Link>
+            </li>
+            {links.map(({ title, link }, key) => {
+                const pathname = link.replace(/^\/dashboard\/?/, "")
+                if (!rbac.includes(pathname)) return null
+                return (
+                    <li
+                        className={merge(
+                            "w-full p-2 border border-transparent rounded hover:text-white hover:border-sky-500 hover:cursor-pointer hover:bg-sky-500",
+                            classNameOption
+                        )}
+                        key={key}
+                    >
+                        <Link href={link}>{title}</Link>
+                    </li>
+                )
+            })}
+        </motion.ul>
+    )
 }
