+++ src/ui/dashboard/contact-people/add-contact-person.tsx
@@ -1,6 +1,5 @@
 "use client"
 import { useEffect, useState, useActionState } from "react"
-import { redirect } from "next/navigation"
 import { Project, Role } from "@prisma/client"
 import { addContactPersonAction } from "@/lib/actions"
 import { AddContactPersonActionState } from "@/lib/@types/types"
@@ -8,7 +7,7 @@ import { getProjectsByStakeHolderId, getRoles } from "@/lib/services"
 import { Form, InputList, Label, SelectGeneric, Submit } from "@/ui/common/form-elements"
 import { ClassNameProps } from "@/lib/@types/props"
 import { merge } from "@halvaradop/ui-core"
-import { getCookieToken } from "@/lib/services/cookies"
+import { getSessionToken } from "@/lib/utils"
 import dataJson from "@/lib/data.json"
 
 const { contactPersonInputs } = dataJson
@@ -24,11 +23,8 @@ export const AddContactPerson = ({ className }: ClassNameProps) => {
 
     useEffect(() => {
         const fetchProjects = async () => {
-            const { ok, data } = await getCookieToken()
-            if (!ok) {
-                return redirect("/dashboard?error=You need to select a stakeholder first")
-            }
-            const response = await getProjectsByStakeHolderId(data.idStakeHolder)
+            const { idStakeHolder } = await getSessionToken()
+            const response = await getProjectsByStakeHolderId(idStakeHolder)
             setProjects(response)
         }
         fetchProjects()
