+++ src/ui/dashboard/index/card.tsx
@@ -1,19 +1,18 @@
-import Image from "next/image"
+import Link from "next/link"
 import { CardDashboardProps } from "@/lib/@types/props"
 import { merge } from "@/lib/utils"
+import { Button } from "@halvaradop/ui-button"
 
-export const CardDashboard = ({ src, alt, title, count, isHover = false }: CardDashboardProps) => {
+export const CardDashboard = ({ title, count, href }: CardDashboardProps) => {
     return (
-        <figure
-            className={merge("p-3 flex items-center justify-evenly gap-x-4 border border-gray-1000 rounded-lg bg-white ", {
-                "hover:cursor-pointer": isHover,
-            })}
+        <article
+            className={merge("py-3 px-4 flex items-start justify-center flex-col border border-gray-1000 rounded-md bg-white")}
         >
-            <Image width={48} height={48} src={src} alt={`${alt} icon`} priority />
-            <figcaption className="flex flex-col">
-                <h2>{title}</h2>
-                <p>{count}</p>
-            </figcaption>
-        </figure>
+            <h2 className="text-lg font-medium">{title}</h2>
+            <p className="mb-2 text-3xl font-bold">{count}</p>
+            <Button className="font-normal border-gray-1000 bg-slate-50 focus-visible:ring-gray-1000" variant="outline" asChild>
+                <Link href={href}>View More</Link>
+            </Button>
+        </article>
     )
 }
