+++ src/ui/dashboard/index/position-resistivities.tsx
@@ -6,7 +6,6 @@ import {
 import { CardDashboard } from "./card"
 import { Roles } from "@/lib/@types/types"
 import { isFalsy } from "@halvaradop/ts-utility-types/validate"
-import positionMeasurementIcon from "@/public/samples.svg"
 
 export const PositionResistivities = async ({ role, id }: { role?: Roles; id?: string }) => {
     const positions =
@@ -15,12 +14,5 @@ export const PositionResistivities = async ({ role, id }: { role?: Roles; id?: s
             : role === "client-admin"
               ? await getPositionResistivitiesByStakeholderId(id!)
               : await getPositionResistivitiesByContactPerson(id!)
-    return (
-        <CardDashboard
-            src={positionMeasurementIcon}
-            alt="position resistivity"
-            title="Position Resistivities"
-            count={positions.length}
-        />
-    )
+    return <CardDashboard href="/dashboard/position-resistivities" title="Position Resistivities" count={positions.length} />
 }
