+++ src/ui/dashboard/index/position-soil-datas.tsx
@@ -2,7 +2,6 @@ import { CardDashboard } from "./card"
 import { Roles } from "@/lib/@types/types"
 import { getPositionSoilDatasByContactPerson, getPositionSoilDatas, getPositionSoilDatasByStakeholderId } from "@/lib/services"
 import { isFalsy } from "@halvaradop/ts-utility-types/validate"
-import soilData from "@/public/soil-data.png"
 
 export const PositionSoilDatas = async ({ role, id }: { role?: Roles; id?: string }) => {
     const soilDatas =
@@ -12,5 +11,5 @@ export const PositionSoilDatas = async ({ role, id }: { role?: Roles; id?: strin
               ? await getPositionSoilDatasByStakeholderId(id!)
               : await getPositionSoilDatasByContactPerson(id!)
 
-    return <CardDashboard src={soilData} alt="position soil data" title="Position soil datas" count={soilDatas.length} />
+    return <CardDashboard href="/dashboard/position-soil-datas" title="Position soil datas" count={soilDatas.length} />
 }
