+++ src/ui/dashboard/index/projects.tsx
@@ -2,7 +2,6 @@ import { getProjects, getProjectsByStakeHolderId, getProjectsByContactPersonId }
 import { CardDashboard } from "./card"
 import { Roles } from "@/lib/@types/types"
 import { isFalsy } from "@halvaradop/ts-utility-types/validate"
-import projectsIcon from "@/public/projects.png"
 
 export const Projects = async ({ role, id }: { role?: Roles; id?: string }) => {
     const projects =
@@ -11,5 +10,5 @@ export const Projects = async ({ role, id }: { role?: Roles; id?: string }) => {
             : role === "client-admin"
               ? await getProjectsByStakeHolderId(id!)
               : await getProjectsByContactPersonId(id!)
-    return <CardDashboard src={projectsIcon} alt="projects" title="Projects" count={projects.length} />
+    return <CardDashboard href="/dashboard/projects" title="Projects" count={projects.length} />
 }
