+++ src/ui/dashboard/index/stakeholders.tsx
@@ -1,8 +1,7 @@
 import { CardDashboard } from "./card"
 import { getStakeholder } from "@/lib/services"
-import stakeHolderIcon from "@/public/stakeholders.png"
 
 export const Stakeholders = async () => {
     const stakeholders = await getStakeholder()
-    return <CardDashboard src={stakeHolderIcon} alt="stakeholders" title="StakeHolders" count={stakeholders.length} />
+    return <CardDashboard href="/dashboard/stake-holders" title="StakeHolders" count={stakeholders.length} />
 }
