+++ src/ui/header/avatar.tsx
@@ -13,7 +13,7 @@ export const Avatar = () => {
     } = session
 
     return (
-        <div className="w-max flex items-center justify-end gap-x-5">
+        <div className="w-full flex items-center justify-end gap-x-5 base:w-auto">
             <div className="w-max hidden items-center gap-x-5 base:flex">
                 {picture ? (
                     <Image
@@ -33,7 +33,10 @@ export const Avatar = () => {
                     </p>
                 </div>
             </div>
-            <Button className="w-auto mt-4 border-white base:m-0" onClick={() => signOut({ redirectTo: "/" })}>
+            <Button
+                className="w-full mt-4 border-sky-500 bg-sky-500 focus-visible:ring-sky-500 base:w-auto base:m-0"
+                onClick={() => signOut({ redirectTo: "/" })}
+            >
                 Log out
             </Button>
         </div>
